style: style children; display children title as h1
