+++ app/routes/_app._index/route.tsx
@@ -1,17 +1,38 @@
+import i18next from '@lib/i18n/index.server';
 import { authenticate } from '@lib/utils/auth.server';
-import { json, redirect, type LoaderFunctionArgs } from '@remix-run/node';
+import {
+  json,
+  redirect,
+  type LoaderFunctionArgs,
+  type MetaFunction,
+} from '@remix-run/node';
 
-export async function loader({ context: { session } }: LoaderFunctionArgs) {
+export function handle() {
+  return { i18n: ['meta', 'home'] };
+}
+
+export const meta: MetaFunction<typeof loader> = ({ data: { title } = {} }) => {
+  return [{ title }];
+};
+
+export async function loader({
+  request,
+  context: { session },
+}: LoaderFunctionArgs) {
   if (!(await authenticate(session))) {
     session.unset('accessToken');
     session.unset('refreshToken');
     throw redirect('/login');
   }
-  return json(null);
+
+  const title = await i18next
+    .getFixedT(request, 'meta')
+    .then((x) => x('home.title'));
+  return json({ title });
 }
 
 function Home() {
-  return <div>Home</div>;
+  return <div>Homepage</div>;
 }
 
 export default Home;
