+++ app/routes/_app/ToolBar.tsx
@@ -18,7 +18,7 @@ export default function ToolBar() {
   ]);
 
   return (
-    <div className="self-center justify-end items-center lg:p-4 flex gap-3">
+    <div className="self-center justify-end items-center flex gap-3">
       <div className="">
         <div className="border-solid border-primary-950 border-[1px] p-2 rounded-full relative bg-primary-100 bg-opacity-50 border-opacity-20">
           <div className="h-2 aspect-square bg-blue-500 right-2 rounded-full absolute" />
